+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fix installation file first read
 - `check` command
+- `remove` and `retreat` commands
 
 ## 1.0.0.pre.0 (2017-04-23)
 - publish, push, add, update